fix: test and lint
