+++ .cspell.json
@@ -1,7 +1,20 @@
 {
   "$schema": "https://raw.githubusercontent.com/streetsidesoftware/cspell/main/cspell.schema.json",
   "version": "0.2",
-  "ignorePaths": ["**/*.json", "**/*.css", "node_modules", "**/*.log", "./src/adapters/supabase/**/**.ts", "/supabase/*", "dist/**", "bun.lockb", "**/*.xml", "**/*.xlsx", "**/*.docx", "**/*.pptx"],
+  "ignorePaths": [
+    "**/*.json",
+    "**/*.css",
+    "node_modules",
+    "**/*.log",
+    "./src/adapters/supabase/**/**.ts",
+    "/supabase/*",
+    "dist/**",
+    "bun.lockb",
+    "**/*.xml",
+    "**/*.xlsx",
+    "**/*.docx",
+    "**/*.pptx"
+  ],
   "useGitignore": true,
   "language": "en",
   "words": [
