+++ tests/log.test.ts
@@ -1,10 +1,9 @@
 import { jest } from "@jest/globals";
-import { Logs } from "@ubiquity-os/ubiquity-os-logger";
+import { createContext } from "./utils";
 import { Context } from "../src/types";
 
 describe("Log post message test", () => {
   it("Should post a waiting message on start", async () => {
-    const addCommentToIssue = jest.fn();
     jest.unstable_mockModule("../src/handlers/ask-llm", () => ({
       askQuestion: jest.fn(() => ({
         answer: "hello",
@@ -13,25 +12,18 @@ describe("Log post message test", () => {
       })),
     }));
     const { processCommentCallback } = await import("../src/handlers/comment-created-callback");
-    const context = {
-      payload: {
-        comment: {
-          user: {
-            type: "User",
-          },
-          body: "/ask hello",
-        },
-      },
-      logger: new Logs("debug"),
-      env: {
-        UBIQUITY_OS_APP_NAME: "UbiquityOS",
-      },
-      commentHandler: {
-        postComment: addCommentToIssue,
-      },
-    } as unknown as Context;
-
+    const context = createContext("hello");
+    context.config.processDriveLinks = false;
     await processCommentCallback(context);
-    expect(addCommentToIssue.mock.calls[0][1]).toMatchObject({ logMessage: { raw: "Thinking..." } });
+    expect(context.commentHandler.postComment).toHaveBeenNthCalledWith(
+      1,
+      expect.anything(),
+      expect.objectContaining({
+        logMessage: expect.objectContaining({
+          raw: expect.stringContaining("Thinking..."),
+        }),
+      }),
+      expect.anything()
+    );
   });
 });
